+++ kernels/openvx-core/c66x/vx_bitwise_target.c
@@ -126,14 +126,14 @@ vx_status VX_CALLBACK tivxKernelBitwiseProcess(
            Assuming valid Roi is same for src0 and src1 images */
         rect = src0_desc->valid_roi;
 
-        src0_addr = (uint8_t *)((uint32_t)src0_desc->mem_ptr[0U].target_ptr +
+        src0_addr = (uint8_t *)((uintptr_t)src0_desc->mem_ptr[0U].target_ptr +
             ownComputePatchOffset(rect.start_x, rect.start_y,
             &src0_desc->imagepatch_addr[0U]));
-        src1_addr = (uint8_t *)((uint32_t)src1_desc->mem_ptr[0U].target_ptr +
+        src1_addr = (uint8_t *)((uintptr_t)src1_desc->mem_ptr[0U].target_ptr +
             ownComputePatchOffset(rect.start_x, rect.start_y,
             &src1_desc->imagepatch_addr[0U]));
         /* TODO: Do we require to move pointer even for destination image */
-        dst_addr = (uint8_t *)((uint32_t)dst_desc->mem_ptr[0U].target_ptr +
+        dst_addr = (uint8_t *)((uintptr_t)dst_desc->mem_ptr[0U].target_ptr +
             ownComputePatchOffset(rect.start_x, rect.start_y,
             &dst_desc->imagepatch_addr[0]));
 
@@ -222,11 +222,11 @@ vx_status VX_CALLBACK tivxKernelBitwiseNotProcess(
            Assuming valid Roi is same for src0 and src1 images */
         rect = src_desc->valid_roi;
 
-        src_addr = (uint8_t *)((uint32_t)src_desc->mem_ptr[0U].target_ptr +
+        src_addr = (uint8_t *)((uintptr_t)src_desc->mem_ptr[0U].target_ptr +
             ownComputePatchOffset(rect.start_x, rect.start_y,
             &src_desc->imagepatch_addr[0U]));
         /* TODO: Do we require to move pointer even for destination image */
-        dst_addr = (uint8_t *)((uint32_t)dst_desc->mem_ptr[0U].target_ptr +
+        dst_addr = (uint8_t *)((uintptr_t)dst_desc->mem_ptr[0U].target_ptr +
             ownComputePatchOffset(rect.start_x, rect.start_y,
             &dst_desc->imagepatch_addr[0]));
