+++ kernels/openvx-core/c66x/vx_gaussian_pyramid_target.c
@@ -116,9 +116,9 @@ static vx_status VX_CALLBACK tivxKernelGsnPmdProcess(
                 dst->mem_ptr[0].mem_type, VX_WRITE_ONLY);
 
             /* Valid rectangle is ignore here */
-            src_addr = (uint8_t *)((uint32_t)src->mem_ptr[0U].target_ptr +
+            src_addr = (uint8_t *)((uintptr_t)src->mem_ptr[0U].target_ptr +
                 ownComputePatchOffset(0, 0, &src->imagepatch_addr[0U]));
-            dst_addr = (uint8_t *)((uint32_t)dst->mem_ptr[0U].target_ptr +
+            dst_addr = (uint8_t *)((uintptr_t)dst->mem_ptr[0U].target_ptr +
                 ownComputePatchOffset(0, 0, &dst->imagepatch_addr[0]));
 
             vxlib_src.dim_x = src->imagepatch_addr[0].dim_x;
@@ -146,7 +146,7 @@ static vx_status VX_CALLBACK tivxKernelGsnPmdProcess(
                 if(pmd->scale == 0.5f)
                 {
                     dst_addr = (uint8_t *)(
-                        (uint32_t)dst->mem_ptr[0U].target_ptr +
+                        (uintptr_t)dst->mem_ptr[0U].target_ptr +
                         ownComputePatchOffset(
                         1u, 1u, &dst->imagepatch_addr[0]));
