+++ kernels/openvx-core/c66x/vx_optical_flow_pyr_lk_target.c
@@ -201,7 +201,7 @@ static vx_status VX_CALLBACK tivxOpticalFlowPyrLk(
                 old_image->mem_size[0], old_image->mem_ptr[0].mem_type,
                 VX_READ_ONLY);
 
-            old_image_addr = (uint8_t *)((uint32_t)old_image->mem_ptr[0U].target_ptr +
+            old_image_addr = (uint8_t *)((uintptr_t)old_image->mem_ptr[0U].target_ptr +
                 ownComputePatchOffset(0, 0, &old_image->imagepatch_addr[0U]));
 
             /* Map New image */
@@ -212,7 +212,7 @@ static vx_status VX_CALLBACK tivxOpticalFlowPyrLk(
                 new_image->mem_size[0], new_image->mem_ptr[0].mem_type,
                 VX_READ_ONLY); // Not sure if this is READ_ONLY
 
-            new_image_addr = (uint8_t *)((uint32_t)new_image->mem_ptr[0U].target_ptr +
+            new_image_addr = (uint8_t *)((uintptr_t)new_image->mem_ptr[0U].target_ptr +
                 ownComputePatchOffset(0, 0, &new_image->imagepatch_addr[0U]));
 
             /* Fetch ROI */
